refactor: Undo reorganization of pip detectable classes
