+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pip/PipInspectorDetectableOptions.java
@@ -20,22 +20,26 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detectable.detectables.pip.pipenv.model;
+package com.synopsys.integration.detectable.detectables.pip;
 
-public class PipFreezeEntry {
-    private final String name;
-    private final String version;
+import java.nio.file.Path;
+import java.util.List;
+import java.util.Optional;
 
-    public PipFreezeEntry(final String name, final String version) {
-        this.name = name;
-        this.version = version;
+public class PipInspectorDetectableOptions {
+    private final String pipProjectName;
+    private final List<Path> requirementsFilePaths;
+
+    public PipInspectorDetectableOptions(final String pipProjectName, final List<Path> requirementsFilePaths) {
+        this.pipProjectName = pipProjectName;
+        this.requirementsFilePaths = requirementsFilePaths;
     }
 
-    public String getName() {
-        return name;
+    public Optional<String> getPipProjectName() {
+        return Optional.ofNullable(pipProjectName);
     }
 
-    public String getVersion() {
-        return version;
+    public List<Path> getRequirementsFilePaths() {
+        return requirementsFilePaths;
     }
 }
