+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pip/PipInspectorExtractor.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detectable.detectables.pip.inspector;
+package com.synopsys.integration.detectable.detectables.pip;
 
 import java.io.File;
 import java.nio.file.Path;
@@ -35,8 +35,8 @@ import com.synopsys.integration.detectable.Extraction;
 import com.synopsys.integration.detectable.detectable.codelocation.CodeLocation;
 import com.synopsys.integration.detectable.detectable.executable.ExecutableRunner;
 import com.synopsys.integration.detectable.detectable.executable.ExecutableRunnerException;
-import com.synopsys.integration.detectable.detectables.pip.inspector.parser.PipInspectorTreeParser;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.model.PipenvResult;
+import com.synopsys.integration.detectable.detectables.pip.parser.PipInspectorTreeParser;
+import com.synopsys.integration.detectable.detectables.pip.model.PipenvResult;
 
 public class PipInspectorExtractor {
     private final ExecutableRunner executableRunner;
