+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pip/PipenvDetectable.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detectable.detectables.pip.pipenv;
+package com.synopsys.integration.detectable.detectables.pip;
 
 import java.io.File;
