+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pip/model/PipenvGraph.java
@@ -20,18 +20,18 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detectable.detectables.pip.pipenv.model;
+package com.synopsys.integration.detectable.detectables.pip.model;
 
 import java.util.List;
 
-public class PipFreeze {
-    private final List<PipFreezeEntry> entries;
+public class PipenvGraph {
+    private final List<PipenvGraphEntry> entries;
 
-    public PipFreeze(final List<PipFreezeEntry> entries) {
+    public PipenvGraph(final List<PipenvGraphEntry> entries) {
         this.entries = entries;
     }
 
-    public List<PipFreezeEntry> getEntries() {
+    public List<PipenvGraphEntry> getEntries() {
         return entries;
     }
 }
