+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pip/model/PipenvResult.java
@@ -20,35 +20,30 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detectable.detectables.pip.pipenv.model;
+package com.synopsys.integration.detectable.detectables.pip.model;
 
-import java.util.List;
+import com.synopsys.integration.detectable.detectable.codelocation.CodeLocation;
 
-import com.google.gson.annotations.SerializedName;
+public class PipenvResult {
+    private final String projectName;
+    private final String projectVersion;
+    private final CodeLocation codeLocation;
 
-public class PipenvGraphEntry {
-    @SerializedName("package_name")
-    private final String name;
-    @SerializedName("installed_version")
-    private final String version;
-    @SerializedName("dependencies")
-    private final List<PipenvGraphDependency> children;
-
-    public PipenvGraphEntry(final String name, final String version, final List<PipenvGraphDependency> children) {
-        this.name = name;
-        this.version = version;
-        this.children = children;
+    public PipenvResult(final String projectName, final String projectVersion, final CodeLocation codeLocation) {
+        this.projectName = projectName;
+        this.projectVersion = projectVersion;
+        this.codeLocation = codeLocation;
     }
 
-    public String getName() {
-        return name;
+    public String getProjectName() {
+        return projectName;
     }
 
-    public String getVersion() {
-        return version;
+    public String getProjectVersion() {
+        return projectVersion;
     }
 
-    public List<PipenvGraphDependency> getChildren() {
-        return children;
+    public CodeLocation getCodeLocation() {
+        return codeLocation;
     }
 }
