+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/pip/parser/PipEnvJsonGraphParser.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detectable.detectables.pip.pipenv.parser;
+package com.synopsys.integration.detectable.detectables.pip.parser;
 
 import java.lang.reflect.Type;
 import java.util.List;
@@ -31,8 +31,8 @@ import org.jetbrains.annotations.NotNull;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.model.PipenvGraph;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.model.PipenvGraphEntry;
+import com.synopsys.integration.detectable.detectables.pip.model.PipenvGraph;
+import com.synopsys.integration.detectable.detectables.pip.model.PipenvGraphEntry;
 
 public class PipEnvJsonGraphParser {
     private static final Type PIPENV_GRAPH_ENTRY_TYPE = new TypeToken<List<PipenvGraphEntry>>() {}.getType();
