+++ detectable/src/main/java/com/synopsys/integration/detectable/factory/DetectableFactory.java
@@ -164,16 +164,16 @@ import com.synopsys.integration.detectable.detectables.pear.parse.PearListParser
 import com.synopsys.integration.detectable.detectables.pear.parse.PearPackageDependenciesParser;
 import com.synopsys.integration.detectable.detectables.pear.parse.PearPackageXmlParser;
 import com.synopsys.integration.detectable.detectables.pear.transform.PearDependencyGraphTransformer;
-import com.synopsys.integration.detectable.detectables.pip.inspector.PipInspectorDetectable;
-import com.synopsys.integration.detectable.detectables.pip.inspector.PipInspectorDetectableOptions;
-import com.synopsys.integration.detectable.detectables.pip.inspector.PipInspectorExtractor;
-import com.synopsys.integration.detectable.detectables.pip.inspector.parser.PipInspectorTreeParser;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.PipenvDetectable;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.PipenvDetectableOptions;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.PipenvExtractor;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.parser.PipEnvJsonGraphParser;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.parser.PipenvFreezeParser;
-import com.synopsys.integration.detectable.detectables.pip.pipenv.parser.PipenvTransformer;
+import com.synopsys.integration.detectable.detectables.pip.PipInspectorDetectable;
+import com.synopsys.integration.detectable.detectables.pip.PipInspectorDetectableOptions;
+import com.synopsys.integration.detectable.detectables.pip.PipInspectorExtractor;
+import com.synopsys.integration.detectable.detectables.pip.parser.PipInspectorTreeParser;
+import com.synopsys.integration.detectable.detectables.pip.PipenvDetectable;
+import com.synopsys.integration.detectable.detectables.pip.PipenvDetectableOptions;
+import com.synopsys.integration.detectable.detectables.pip.PipenvExtractor;
+import com.synopsys.integration.detectable.detectables.pip.parser.PipEnvJsonGraphParser;
+import com.synopsys.integration.detectable.detectables.pip.parser.PipenvFreezeParser;
+import com.synopsys.integration.detectable.detectables.pip.parser.PipenvTransformer;
 import com.synopsys.integration.detectable.detectables.pip.poetry.PoetryDetectable;
 import com.synopsys.integration.detectable.detectables.pip.poetry.PoetryExtractor;
 import com.synopsys.integration.detectable.detectables.pip.poetry.parser.PoetryLockParser;
